And some more stock tests

+++ src/ThreeFourteen.Finnhub.Client/Model/Candle.cs
@@ -5,10 +5,10 @@ namespace ThreeFourteen.Finnhub.Client.Model
     public class Candle
     {
         public DateTime Timestamp { get; set; }
-        public double High { get; set; }
-        public double Low { get; set; }
-        public double Open { get; set; }
-        public double Close { get; set; }
-        public int Volume { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
+        public decimal Open { get; set; }
+        public decimal Close { get; set; }
+        public long Volume { get; set; }
     }
 }
