+++ src/ThreeFourteen.Finnhub.Client/Model/Earnings.cs
@@ -6,10 +6,10 @@ namespace ThreeFourteen.Finnhub.Client.Model
     public class Earnings
     {
         [JsonProperty("actual")]
-        public double Actual { get; set; }
+        public decimal Actual { get; set; }
 
         [JsonProperty("estimate")]
-        public double Estimate { get; set; }
+        public decimal Estimate { get; set; }
 
         [JsonProperty("period")]
         public DateTime Period { get; set; }
